requirements.txt
Django==4.2.2
django-cors-headers==4.2.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
djlint==1.34.0
mysqlclient==2.2.0
Pillow==9.5.0